% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_orgas_wrapper.R
\name{clean_orgas_wrapper}
\alias{clean_orgas_wrapper}
\title{Organisationen bereinigen und abgleichen}
\usage{
clean_orgas_wrapper(scraped_df, gerit_df)
}
\arguments{
\item{scraped_df}{Ein Dataframe mit gescrapten Organisationsdaten.}

\item{gerit_df}{Ein Dataframe mit GERIT-Organisationsdaten.}
}
\value{
Ein Dataframe, das die gematchten Organisationen enthält, inklusive Matching-Score und Matching-Art.
}
\description{
Diese Funktion nimmt zwei Dataframes (scraped_df und gerit_df), konvertiert sie in pandas-DataFrames,
führt einen Abgleich der Organisationen mittels einer Python-Funktion durch und gibt das gematchte Dataframe zurück.
}
\details{
Die Funktion ruft eine Python-Funktion auf, um die Organisationen abzugleichen. Anschließend werden nur die relevanten
Variablen übernommen und das Ergebnis mit dem ursprünglichen Dataframe gemerged.
}
\examples{
\dontrun{
result <- clean_orga(scraped_df, gerit_df)
}
}
