% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_and_join_classified_data.r
\name{merge_and_join_classified_data}
\alias{merge_and_join_classified_data}
\title{Merge old + new classified data and join fixed classification vars to base data}
\usage{
merge_and_join_classified_data(
  raw_data,
  db_data_path,
  new_classified_data,
  key_vars = c("titel", "nummer")
)
}
\arguments{
\item{raw_data}{Ein data.frame mit allen Rohdaten.}

\item{db_data_path}{Pfad zur bestehenden RDS-Datei mit bereits klassifizierten Daten.}

\item{new_classified_data}{Ein data.frame mit neu klassifizierten Daten.}

\item{key_vars}{Ein Vektor mit Spaltennamen für den Join (Default: c("titel","nummer")).}
}
\value{
Ein data.frame mit allen Rohdaten inkl. den Klassifizierungsspalten.
}
\description{
Lädt bestehende Klassifizierungsdaten, kombiniert sie mit neuen Klassifizierungen
(ohne Duplikate), reduziert auf die festen Klassifizierungsspalten und merged
das Ergebnis per left_join an die Rohdaten.
}
